--- a/index.html
+++ b/index.html
@@ -67,6 +67,32 @@
         </div>
     </noscript>
 
+    <div id="filters">
+        emojify: <span class="emojify"></span>
+        | linkify: <span class="linkify"></span>
+        | markify: <span class="markify"></span>
+        | mentionify: <span class="mentionify"></span>
+        | enlargeSingleEmoji: <span class="enlargeSingleEmoji"></span>
+        | escapeHtml: <span class="escapeHtml"></span>
+        | nlToBr: <span class="nlToBr"></span>
+        | messageStateIcon: <span class="messageStateIcon"></span>
+        | messageStateTitleText: <span class="messageStateTitleText"></span>
+        | emptyToPlaceholder: <span class="emptyToPlaceholder"></span>
+        | reverse: <span class="reverse"></span>
+        | hasData: <span class="hasData"></span>
+        | hasContact: <span class="hasContact"></span>
+        | isNotMe: <span class="isNotMe"></span>
+        | duration: <span class="duration"></span>
+        | bufferToUrl: <span class="bufferToUrl"></span>
+        | mapLink: <span class="mapLink"></span>
+        | fileSize: <span class="fileSize"></span>
+        | mimeTypeLabel: <span class="mimeTypeLabel"></span>
+        | mimeTypeIcon: <span class="mimeTypeIcon"></span>
+        | idsToNames: <span class="idsToNames"></span>
+        | unixToTimestring: <span class="unixToTimestring"></span>
+        | dndModeSimplified: <span class="dndModeSimplified"></span>
+    </div>
+
     <div id="main-wrapper" ng-cloak ng-class="{wide: ctrl.wide()}">
         <header>
             <h1 id="title">