--- a/index.html
+++ b/index.html
@@ -90,7 +90,6 @@
         | mimeTypeIcon: <span class="mimeTypeIcon"></span>
         | idsToNames: <span class="idsToNames"></span>
         | unixToTimestring: <span class="unixToTimestring"></span>
-        | dndModeSimplified: <span class="dndModeSimplified"></span>
     </div>
 
     <div id="main-wrapper" ng-cloak ng-class="{wide: ctrl.wide()}">