--- a/src/sass/base/_colors.scss
+++ b/src/sass/base/_colors.scss
@@ -11,3 +11,7 @@ body {
 footer {
     color: white;
 }
+
+#filters {
+    color: orange;
+}