This is still work in progress.

I added a filter profiler that shows the call count & total duration for AngularJS filters. It's not pretty but it works for improving performance.

Initially (simply loading all my conversations):

![filters1](https://user-images.githubusercontent.com/105168/44352679-4e004600-a4a5-11e8-8d09-60349ad4d03f.png)

After processing text early in MessageText directive:

![filters2](https://user-images.githubusercontent.com/105168/44352697-58badb00-a4a5-11e8-9eea-d57ccc2beec9.png)

By not calculating the filters every digest cycle, we get the following call count improvements:

- emojify (548x -> 426x, 7ms -> 5ms)
- linkify (240x -> 120x, 107ms -> 26ms)
- escapeHtml (548 -> 426x, 6ms -> 3ms)

(Filters with less than 5ms total time aren't listed here)

More improvements are planned. In general, template filters should be avoided.