@dbrgn 2018-10-25
Closing this for now. The important changes from here are now in master, we can always continue working on this branch later.